requirements.txt
# AI Quality Assurance Orchestrator - Dependencies

# Core AI Framework
crewai>=0.152.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.8.0
# openai>=1.3.0  # Optional - only if you want to use OpenAI models

# Web Framework & UI
streamlit>=1.29.0
plotly>=5.17.0
dash>=2.14.1

# Data Processing & Analysis
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0

# Testing Frameworks
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# AI Validation & Evaluation Metrics
sentence-transformers>=2.2.0
evaluate>=0.4.0
nltk>=3.8.0
textstat>=0.7.0

# Code Analysis & Security (Optional - install separately if needed)
# bandit>=1.7.0
# safety>=2.3.0

# Performance Testing (Optional - install separately if needed)  
# locust>=2.17.0

# Advanced AI Evaluation (Optional - for production use)
# ragas>=0.1.0
# deepeval>=0.20.0
# bert-score>=0.3.0

# Utilities
typing-extensions>=4.8.0
rich>=13.7.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Development (Optional)
# black>=23.11.0
# flake8>=6.1.0